Verallgemeinern der ServiceXml
