+++ src/main/java/Kaufvertrag/dataLayer/dataAccessObjects/XML/ServiceXML.java
@@ -22,27 +22,15 @@ import java.util.List;
 
 public class ServiceXML
 {
-    private static final String WARE_XML = "LF5or8-Gruppenarbeit/src/main/java/Kaufvertrag/dataLayer/dataAccessObjects/XML/wareTest.xml"; // Adjust the path as needed
-
-    public static void main(String[] args)
-    {
-        List<String> besonderheiten = List.of("Ist toll");
-        List<String> maengel = List.of("ist aber auch kaputt");
-
-        Ware test = new Ware("test", "test", 20, besonderheiten, maengel);
-        test.setId(10);
-        //write(test);
-    }
-
-    public static Document read()
+    public static Document read(String path)
     {
         Document doc = null;
         try
         {
-            File xmlFile = new File(WARE_XML);
+            File xmlFile = new File(path);
             if (!xmlFile.exists())
             {
-                System.out.println("The file " + WARE_XML + " does not exist.");
+                System.out.println("The file " + path + " does not exist.");
                 return null;
             }
 
@@ -60,6 +48,7 @@ public class ServiceXML
         return doc;
     }
 
+    // For testing purposes only
     private static void printDocument(Document doc) throws TransformerException {
         TransformerFactory tf = TransformerFactory.newInstance();
         Transformer transformer = tf.newTransformer();
@@ -70,7 +59,7 @@ public class ServiceXML
         transformer.transform(source, console);
     }
 
-    public static void write(IWare ware)
+    public static void write(Element domElementToWrite, String path)
     {
         try
         {
@@ -79,25 +68,23 @@ public class ServiceXML
             DocumentBuilder dBuilder = dbFactory.newDocumentBuilder();
             Document doc = dBuilder.newDocument();
 
-            File xmlFile = new File(WARE_XML);
+            File xmlFile = new File(path);
 
             if (!xmlFile.exists())
             {
                 boolean isFileCreated = xmlFile.createNewFile();
                 if (!isFileCreated)
                 {
-                    throw new IOException("Failed to create new file: " + WARE_XML);
+                    throw new IOException("Failed to create new file: " + path);
                 }
             }
 
             // Create root element
-            Element rootElement = doc.createElement("wareList");
+            Element rootElement = doc.createElement("rootList");
             rootElement.setAttribute("xmlns:w3s", "https://www.w3schools.com");
             doc.appendChild(rootElement);
 
-            // Convert Ware object to XML Element
-            Element wareElement = createWareElement(doc, ware);
-            rootElement.appendChild(wareElement);
+            rootElement.appendChild(domElementToInsert);
 
             // Write the content into an XML file
             TransformerFactory transformerFactory = TransformerFactory.newInstance();
@@ -105,7 +92,7 @@ public class ServiceXML
             transformer.setOutputProperty(OutputKeys.INDENT, "yes");
             transformer.setOutputProperty("{http://xml.apache.org/xslt}indent-amount", "2");
             DOMSource source = new DOMSource(doc);
-            StreamResult result = new StreamResult(new File(WARE_XML));
+            StreamResult result = new StreamResult(new File(path));
 
             transformer.transform(source, result);
 
@@ -124,41 +111,5 @@ public class ServiceXML
         }
     }
 
-    private static Element createWareElement(Document doc, IWare ware)
-    {
-        String prefix = "w3s:";
-        Element wareElement = doc.createElementNS("https://www.w3schools.com", prefix + "ware");
 
-        Element idElement = doc.createElementNS("https://www.w3schools.com", prefix + "id");
-        idElement.appendChild(doc.createTextNode(String.valueOf(ware.getId())));
-        wareElement.appendChild(idElement);
-
-        Element bezeichnungElement = doc.createElementNS("https://www.w3schools.com", prefix + "bezeichnung");
-        bezeichnungElement.appendChild(doc.createTextNode(ware.getBezeichnung()));
-        wareElement.appendChild(bezeichnungElement);
-
-        Element beschreibungElement = doc.createElementNS("https://www.w3schools.com", prefix + "beschreibung");
-        beschreibungElement.appendChild(doc.createTextNode(ware.getBeschreibung()));
-        wareElement.appendChild(beschreibungElement);
-
-        Element preisElement = doc.createElementNS("https://www.w3schools.com", prefix + "preis");
-        preisElement.appendChild(doc.createTextNode(String.valueOf(ware.getPreis())));
-        wareElement.appendChild(preisElement);
-
-        for (String besonderheit : ware.getBesonderheiten())
-        {
-            Element besonderheitElement = doc.createElementNS("https://www.w3schools.com", prefix + "besonderheit");
-            besonderheitElement.appendChild(doc.createTextNode(besonderheit));
-            wareElement.appendChild(besonderheitElement);
-        }
-
-        for (String mangel : ware.getMaengel())
-        {
-            Element mangelElement = doc.createElementNS("https://www.w3schools.com", prefix + "mangel");
-            mangelElement.appendChild(doc.createTextNode(mangel));
-            wareElement.appendChild(mangelElement);
-        }
-
-        return wareElement;
-    }
 }
